Correctly escape numerical options and improve SQL construction
The option-getting method already had a way to return integer
values but it wasn't being used, so this uses it. It also adds
a cast in the SQL string construction to make it even more
clear that the value can't inject arbitrary strings.

Change-Id: I15b883bf9b09c2b9aff43d0cace5f6a1525872b5
(cherry picked from commit 646012354a0196e663265cdeec98e150b9ccecfa)